+++ utils/roadie-backstage-entity-validator/src/__mocks__/fs.js
@@ -1,2 +1,3 @@
 const fs = require('memfs');
+
 module.exports = fs;
