+++ utils/roadie-backstage-entity-validator/src/validator.js
@@ -19,7 +19,7 @@ import { templateEntityV1beta3Validator } from '@backstage/plugin-scaffolder-com
 import annotationSchema from './schemas/annotations.schema.json';
 import Ajv from 'ajv';
 import ajvFormats from 'ajv-formats';
-import path from 'path';
+import { relativeSpaceValidation } from './relativeSpaceValidation';
 
 const ajv = new Ajv({ verbose: true });
 ajvFormats(ajv);
@@ -120,65 +120,6 @@ export const validate = async (fileContents, verbose = true) => {
       console.log('Entity Schema policies validated\n');
       responses.forEach(it => console.log(yaml.dump(it)));
     }
-  } catch (e) {
-    throw new Error(e);
-  }
-};
-
-const relativeSpaceValidation = async (fileContents, filePath, verbose) => {
-  const fileExists = filePath => {
-    let flag = true;
-    try {
-      fs.accessSync(filePath, fs.constants.F_OK);
-    } catch (e) {
-      flag = false;
-    }
-    return flag;
-  };
-
-  const validateTechDocs = async (data, filePath) => {
-    console.log(filePath, '######');
-    if (
-      !data?.metadata?.annotations ||
-      !data?.metadata?.annotations['backstage.io/techdocs-ref']
-    ) {
-      return;
-    }
-    const techDocsAnnotation =
-      data.metadata.annotations['backstage.io/techdocs-ref'];
-    if (!techDocsAnnotation.includes('dir')) {
-      return;
-    }
-
-    const mkdocsYamlPath = path.join(
-      path.dirname(filePath),
-      techDocsAnnotation.split(':')[1],
-      'mkdocs.yaml',
-    );
-    const mkdocsYmlPath = path.join(
-      path.dirname(filePath),
-      techDocsAnnotation.split(':')[1],
-      'mkdocs.yml',
-    );
-
-    if (!fileExists(mkdocsYamlPath) & !fileExists(mkdocsYmlPath)) {
-      throw new Error(
-        `Techdocs annotation specifies "dir" but file under ${mkdocsYamlPath}|${mkdocsYmlPath} not found`,
-      );
-    }
-    return;
-  };
-
-  try {
-    const data = yaml.loadAll(fileContents);
-    if (verbose) {
-      console.log('Validating locally dependant catalog contents');
-    }
-    await Promise.all(
-      data.map(async it => {
-        await validateTechDocs(it, filePath);
-      }),
-    );
   } catch (e) {
     throw new Error(e);
   }
